TropicalCyclonePotentialIntensity

Documentation for TropicalCyclonePotentialIntensity.

TropicalCyclonePotentialIntensity.find_root_newton_raphson

TropicalCyclonePotentialIntensity.get_buoyancy_of_lifted_parcel

TropicalCyclonePotentialIntensity.get_lifted_condensation_level

TropicalCyclonePotentialIntensity.get_minimum_pressure_of_tropical_cyclone

TropicalCyclonePotentialIntensity.get_mixing_ratio

TropicalCyclonePotentialIntensity.get_partial_vapor_pressure

TropicalCyclonePotentialIntensity.get_potential_temperature

TropicalCyclonePotentialIntensity.get_saturation_vapor_pressure

TropicalCyclonePotentialIntensity.get_specific_entropy

TropicalCyclonePotentialIntensity.get_specific_entropy_emanuel

TropicalCyclonePotentialIntensity.get_virtual_temperature

TropicalCyclonePotentialIntensity.get_virtual_temperature

TropicalCyclonePotentialIntensity.mixing_ratio_to_specific_humidity

TropicalCyclonePotentialIntensity.specific_humidity_to_mixing_ratio

TropicalCyclonePotentialIntensity.∂specific_entropy_∂temp

TropicalCyclonePotentialIntensity.find_root_newton_raphson — Method

approximate_t_newton_raphson(tguess,penv,target_entropy)

Given a parcel's total entropy sparcel and a pressure level P, use the Newton-Raphson method to find T⋆ such that s(T⋆,P) = sparcel. It assumes reversible moist process. Total entropy of the parcel before condensation: s = (cpd + rtcl)ln(T) - Rdln(pd) + Lvr/T - rRvln(H) : rt is total water (vapor plus liquid)

TropicalCyclonePotentialIntensity.get_buoyancy_of_lifted_parcel — Function

get_buoyancy_of_lifted_parcel(tparcel,rparcel,pparcel,t,r,p,ptop=50)

TropicalCyclonePotentialIntensity.get_lifted_condensation_level — Method

get_lifted_condensation_level(temperature,relative_humidity,pressure)

Receive temperature in Kelvin, relative humidity (unitless) and pressure (hPa) and compute the lifted condensation level based on Emanuel's E94 "calcsound.f" code at http://texmex.mit.edu/pub/emanuel/BOOK/

TropicalCyclonePotentialIntensity.get_minimum_pressure_of_tropical_cyclone — Method

get_minimum_pressure_of_tropical_cyclone(sea_surface_temp,sea_surface_pressure, pressure, temperature, mixing_ratio; ckovercd = 0.9, reversible_ascent=1, dissipative_heating = true)

temperatures in kelvin

TropicalCyclonePotentialIntensity.get_mixing_ratio — Method

get_mixing_ratio(water_vapor_partial_pressure,env_pressure)

Receive a water vapor mixing ratio (unitless g/g) and environmental pressure and compute the partial pressure of water vapor in the same units as the incoming pressure.

TropicalCyclonePotentialIntensity.get_partial_vapor_pressure — Method

get_partial_vapor_pressure(mixing_ratio,pressure)

Receive a water vapor mixing ratio (unitless g/g) and environmental pressure and compute the partial pressure of water vapor in the same units as the input pressure.

TropicalCyclonePotentialIntensity.get_potential_temperature — Method

get_potential_temperature(temperature, pressure, reference_pressure)

Compute potential temperature from temperature and pressure.

TropicalCyclonePotentialIntensity.get_saturation_vapor_pressure — Method

get_saturation_vapor_pressure(T)

Receive temperature T in Kelvin and compute the saturation vapor pressure in hPa from the August-Roche-Magnus formula that approximates the solution to the Clausius-Clapeyron relationship (Wikipedia contributors. (2020, December 19). Clausius–Clapeyron relation. In Wikipedia, The Free Encyclopedia. Retrieved 06:57, December 20, 2020, from https://en.wikipedia.org/w/index.php?title=Clausius%E2%80%93Clapeyron_relation&oldid=995159175)

TropicalCyclonePotentialIntensity.get_specific_entropy — Method

get_specific_entropy(temperature,mixing_ratio,pressure)

Receive temperature in Kelvin, water vapor mixing ratio (unitless g/g) and pressure (hPa) and compute the specific entropy of a parcel using equation in Emmanuel's (E94, EQN. 4.5.9)

TropicalCyclonePotentialIntensity.get_specific_entropy_emanuel — Method

get_specific_entropy_emanuel(temperature,mixing_ratio,pressure)

Receive temperature in Kelvin, water vapor mixing ratio (unitless g/g) and pressure (hPa) and compute the specific entropy of a parcel using equation in Emmanuel's (E94, EQN. 4.5.9)

TropicalCyclonePotentialIntensity.get_virtual_temperature — Method

get_virtual_temperature(temperature,mixing_ratio_total_water,mixing_ratio_water_vapor)

Receive temperature (K) and mixing ratios of total water and water vapor (unitless g/g) and compute the virtual temperature

TropicalCyclonePotentialIntensity.get_virtual_temperature — Method

get_virtual_temperature(temperature, specific_humidity)

Compute virtual temperature from temperature and specific humidity.

TropicalCyclonePotentialIntensity.mixing_ratio_to_specific_humidity — Method

mixing_ratio_to_specific_humidity(mixing_ratio)

Take a mixing ratio (unitless g/g) and return a specific humidity

TropicalCyclonePotentialIntensity.specific_humidity_to_mixing_ratio — Method

specific_humidity_to_mixing_ratio(specific_humidity)

Take a specific humidity (unitless g/g) and return a mixing ratio

TropicalCyclonePotentialIntensity.∂specific_entropy_∂temp — Method